\name{cdoapi-package}
\alias{cdoapi-package}
\alias{cdoapi}
\docType{package}
\title{
\packageTitle{cdoapi}
}
\description{
\packageDescription{cdoapi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cdoapi}
\packageIndices{cdoapi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cdoapi}

Maintainer: \packageMaintainer{cdoapi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
