% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdoapi.R
\docType{package}
\name{cdoapi}
\alias{cdoapi}
\alias{cdoapi-package}
\title{cdoapi: R Interface for the NOAA Climate Data Online web API}
\description{
This package provides an R interface for the climate and weather
data from the NOAA Climate Data Online web API
}
\section{Objects}{

PLACEHOLDER
}

\section{Authors}{

Hariprasath Govindarajan, Hector Plata and Julius Kittler
}

\section{Maintainer}{

hargo729@student.liu.se
}

\section{Reference}{

API: https://www.ncdc.noaa.gov/cdo-web/webservices/v2
}

