% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cities.R
\name{get_cities}
\alias{get_cities}
\title{Get information about available cities}
\usage{
get_cities()
}
\value{
Data frame containing information about the available cities.
}
\description{
\code{get_cities} is a function that lists all the available cities for 
which the data can be queried with \code{get_weatherdata}
}
\examples{
\dontrun{

# First get your API-token here: https://www.ncdc.noaa.gov/cdo-web/token
# Then, set it as system variable (as character variable):
Sys.setenv("NOAA_TOKEN" = "YOUR_TOKEN_GOES_IN_HERE")

cities = get_cities()
head(cities)

head(cities[which(cities$country == "SW"), ]) # Cities in Sweden
head(cities[which(cities$country == "GM"), ]) # Cities in Germany
}

}
\references{
API Reference: https://www.ncdc.noaa.gov/cdo-web/webservices/v2#locations
}
