% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get information about available cities}
\usage{
get_locations()
}
\value{
Data frame containing information about the available cities.
}
\description{
\code{get_locations} is a function that lists all the available cities for 
which the data can be queried with \code{get_weatherdata}
}
\examples{

locations = get_locations()
head(locations)

head(locations[which(locations$country == "SW"), ]) # Locations in Sweden
head(locations[which(locations$country == "GM"), ]) # Locations in Germany

}
