% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_weatherdata.R
\name{simplify_weatherdata}
\alias{simplify_weatherdata}
\title{Simplify weather data for multi-station cities}
\usage{
simplify_weatherdata(weatherdata)
}
\arguments{
\item{weatherdata}{Data frame containing the result of \code{\link{get_weatherdata}}
function}
}
\value{
Modified data frame containing one row for each day; 2 columns containing date 
and mean weather data value for that date.
}
\description{
For multi-station cities, aggregates numeric values of weather data, calculates 
their mean value for each day and simplifies the columns to an easy-to-use daily
data format.
}
\examples{
\dontrun{
# First get your API-token here: https://www.ncdc.noaa.gov/cdo-web/token
# Then, set it as system variable (as character variable):
Sys.setenv("NOAA_TOKEN" = "YOUR_TOKEN_GOES_IN_HERE")

# Get weatherdata for Berlin (only September)
weatherdata = get_weatherdata("TAVG", "GM000001", "2018-09-01", "2018-09-23")

# Since there are 2 stations with weather data for Berlin, we modify the data
weatherdata_simplified = simplify_weatherdata(weatherdata)
}

}
