% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cities.R
\name{get_cities}
\alias{get_cities}
\title{Get information about available cities}
\usage{
get_cities()
}
\value{
Data frame containing information about all the cities for which weather 
data is available and the time period of the data. The following columns 
are returned:
\itemize{
  \item mindate - minimum date from which data is available
  \item maxdate - maximum date upto which data is available
  \item datacoverage - measure of data availability in the time period
  \item id - city ID (required for using \code{get_weatherdata})
  \item country - country ID
  \item City - city name
}
}
\description{
\code{get_cities} is a function that lists all the available cities for 
which the data can be fetched with \code{get_weatherdata} along with 
their data availability details.
}
\examples{
\dontrun{

# First get your API-token here: https://www.ncdc.noaa.gov/cdo-web/token
# Then, set it as system variable (as character variable):
Sys.setenv("NOAA_TOKEN" = "YOUR_TOKEN_GOES_IN_HERE")

cities = get_cities()
head(cities)

head(cities[which(cities$country == "SW"), ]) # Cities in Sweden
head(cities[which(cities$country == "GM"), ]) # Cities in Germany
}

}
\references{
API Reference: \url{https://www.ncdc.noaa.gov/cdo-web/webservices/v2#locations}
}
