% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weatherdata.R
\name{get_weatherdata}
\alias{get_weatherdata}
\title{Get weatherdata for specific time period and cities}
\usage{
get_weatherdata(data_type, city_ID, start_date, end_date)
}
\arguments{
\item{data_type}{Character}

\item{city_ID}{Character}

\item{start_date}{Character, valid date}

\item{end_date}{Character, valid date}
}
\value{

}
\description{
get_weatherdata allows the user to get weatherdata for a specified
}
\details{
a) data_type (e.g. TAVG, TMIN, TMAX, SNOW, PRCP, 
   others: https://docs.opendata.aws/noaa-ghcn-pds/readme.html)
b) city_ID (which should be taken from the output of get_locations)
c) start_date (in the format "YYYY-MM-DD")
d) end_date (in the format "YYYY-MM-DD")
}
\examples{
# Get location data to select the city_IDs we are interested in
locations = get_locations()

# Get weatherdata for Linköping (only September)
head(locations[which(locations$country == "SW"), ]) # Pick SW000006 (Linköping)
weatherdata = get_weatherdata("TAVG", "SW000006", "2018-09-01", "2018-09-23")

# Get weatherdata for Berlin (only September)
head(locations[which(locations$country == "GM"), ]) # Pick GM000001 (Berlin)
weatherdata = get_weatherdata("TAVG", "GM000001", "2018-09-01", "2018-09-23")

}
